requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
apscheduler>=3.10.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
websockets>=12.0
aiofiles>=23.0.0
jinja2>=3.1.0
email-validator>=2.1.0
pyotp>=2.9.0
user-agents>=2.2.0
slowapi>=0.1.9
